+++ lib/Conversion/HipToLLVM/CMakeLists.txt
@@ -9,6 +9,7 @@ add_library(HipToLLVM STATIC
   ConvLowering.cpp
   MatmulLowering.cpp
   ElementwiseLowering.cpp
+  PowerLowering.cpp
   ActivationLowering.cpp
   NormLowering.cpp
   GatherLowering.cpp
