+++ lib/Conversion/OnnxToHip/OnnxToHip.cpp
@@ -367,6 +367,7 @@ static mlir::LogicalResult convertComputeOps(mlir::func::FuncOp funcOp,
   populateMatMulConversionPatterns(patterns, ctx);
   populateTransposeConversionPatterns(patterns, ctx);
   populateElementwiseConversionPatterns(patterns, ctx);
+  populatePowerConversionPatterns(patterns, ctx);
   populateActivationConversionPatterns(patterns, ctx);
   populateCastConversionPatterns(patterns, ctx);
   populateReduceSumConversionPatterns(patterns, ctx);
